+++ src/main/java/org/jboss/tools/intellij/openshift/utils/odo/ComponentTypesDeserializer.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Red Hat, Inc.
+ * Copyright (c) 2019 Red Hat, Inc.
  * Distributed under license by Red Hat, Inc. All rights reserved.
  * This program is made available under the terms of the
  * Eclipse Public License v2.0 which accompanies this distribution,
@@ -8,33 +8,34 @@
  * Contributors:
  * Red Hat, Inc. - initial API and implementation
  ******************************************************************************/
-package org.jboss.tools.intellij.openshift.utils.xml;
+package org.jboss.tools.intellij.openshift.utils.odo;
 
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.deser.std.StdNodeBasedDeserializer;
 import com.fasterxml.jackson.databind.type.TypeFactory;
-import org.jboss.tools.intellij.openshift.utils.odo.ServiceTemplate;
+import org.jboss.tools.intellij.openshift.utils.odo.ComponentType;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
-public class ServiceTemplatesDeserializer extends StdNodeBasedDeserializer<List<ServiceTemplate>> {
-    public ServiceTemplatesDeserializer() {
-        super(TypeFactory.defaultInstance().constructCollectionType(List.class, ServiceTemplate.class));
+public class ComponentTypesDeserializer extends StdNodeBasedDeserializer<List<ComponentType>> {
+    public ComponentTypesDeserializer() {
+        super(TypeFactory.defaultInstance().constructCollectionType(List.class, ComponentType.class));
     }
-
     @Override
-    public List<ServiceTemplate> convert(JsonNode root, DeserializationContext ctxt) throws IOException {
-        List<ServiceTemplate> result = new ArrayList<>();
+    public List<ComponentType> convert(JsonNode root, DeserializationContext ctxt) throws IOException {
+        List<ComponentType> result = new ArrayList<>();
         JsonNode items = root.get("items");
         if (items != null) {
-            for (JsonNode item : items) {
+            for (Iterator<JsonNode> it = items.iterator(); it.hasNext(); ) {
+                JsonNode item = it.next();
                 String name = item.get("metadata").get("name").asText();
-                //TODO manage plan lists.
-                String plan = item.get("spec").get("planList").get(0).asText();
-                result.add(new ServiceTemplate() {
+                List<String> versions = new ArrayList<>();
+                item.get("spec").get("nonHiddenTags").forEach(node -> versions.add(node.textValue()));
+                result.add(new ComponentType() {
 
                     @Override
                     public String getName() {
@@ -42,8 +43,8 @@ public class ServiceTemplatesDeserializer extends StdNodeBasedDeserializer<List<
                     }
 
                     @Override
-                    public String getPlan() {
-                        return plan;
+                    public String[] getVersions() {
+                        return versions.toArray(new String[versions.size()]);
                     }
                 });
             }
